pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pulse-kraken"
version = "0.1.0"
description = "Kraken adapter for PULSE Protocol — trade with PULSE messages"
license = "Apache-2.0"
requires-python = ">=3.8"
authors = [
    {name = "PULSE Protocol Team", email = "pulse@protocol.org"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Office/Business :: Financial :: Investment",
]
dependencies = [
    "pulse-protocol>=0.5.0",
    "requests>=2.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.urls]
Homepage = "https://github.com/pulseprotocolorg-cyber/pulse-kraken"
Repository = "https://github.com/pulseprotocolorg-cyber/pulse-kraken"

[tool.setuptools.packages.find]
include = ["pulse_kraken*"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.black]
line-length = 100
